+++ HolubLife/src/com/holub/life/Life.java
@@ -1,17 +1,15 @@
 package com.holub.life;
 
 import java.awt.*;
-import java.awt.event.ComponentAdapter;
-import java.awt.event.ComponentEvent;
 import javax.swing.*;
 
 import com.holub.life.controller.Clock;
 import com.holub.life.controller.Universe;
 import com.holub.life.model.DummyCell;
+import com.holub.life.universe_settings.SampleSetting;
+import com.holub.life.universe_settings.SettingA;
 import com.holub.life.view.UniverseView;
-import com.holub.life.universe_settings.DummySettings;
 import com.holub.ui.MenuSite;
-import com.sun.org.apache.bcel.internal.generic.DUP;
 
 /*******************************************************************
  * An implemenation of Conway's Game of Life.
@@ -42,10 +40,11 @@ public final class Life extends JFrame
 
 		setDefaultCloseOperation	( EXIT_ON_CLOSE 		);
 		getContentPane().setLayout	( new BorderLayout()	);
-
-		DummySettings dummy = new DummySettings();
-		dummy.establish();
 		getContentPane().add( universeview, BorderLayout.CENTER); //{=life.java.install}
+
+		//리스너를 더해요
+		SampleSetting now_setting = new SettingA();
+		now_setting.establish();
 
 		//pack()은 프레임내에 서브컴포넌트들의 레이아웃과 Prefered Size에
 		//맞도록 윈도우의 사이즈를 맞추는 작업이다. -> 딱히 건들 필요 없다
