+++ HolubLife/src/com/holub/life/universe_settings/SampleSetting.java
@@ -1,0 +1,11 @@
+package com.holub.life.universe_settings;
+
+public abstract class SampleSetting {
+    public void establish(){
+        view_setting();
+        menu_setting();
+    }
+
+    abstract void view_setting();
+    abstract void menu_setting();
+}
